requirements.txt
flask
openai
requests
bs4
markdown2
datetime
gitpython